+++ src/options/SectionSnippets/SectionSnippets.vue
@@ -1,0 +1,18 @@
+<script>
+  import SnippetsIntro from '../SnippetsIntro';
+  import Snippets from '../Snippets/Snippets';
+
+  export default {
+    components: {
+     SnippetsIntro,
+     Snippets,
+    },
+  }
+</script>
+
+<template>
+  <div>
+    <SnippetsIntro />
+    <Snippets />
+  </div>
+</template>