+++ src/options/constants/RouteNames.js
@@ -1,0 +1,4 @@
+export default {
+  General: `general`,
+  Snippets: `snippets`,
+}