+++ src/popup/main.js
@@ -1,9 +1,19 @@
 import Vue from 'vue'
 import Popup from './Popup'
 import '@/global.css'
+import { getOptions } from '@/lib/options'
 
-/* eslint-disable no-new */
-new Vue({
-  el: `#app`,
-  render: h => h(Popup),
-})
+(async () => {
+  const options = await getOptions()
+  console.log(options)
+
+  /* eslint-disable no-new */
+  new Vue({
+    el: `#app`,
+    render: h => h(Popup, {
+      props: {
+        options,
+      },
+    }),
+  })
+})()